Added ability to update size classes for species. Doesn't seem to update the value displayed in the table. Will probably replace it with another JTable.
The size class updating window seems to be used for all of the arraylist
cell editors. Need to read up about this. Also need to add sanity
checking for size classes.

Also seem to be having a but of class-itis. 

Still need to figure out how to grow a colony.